requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
attrs==26.1.0
backtrader==1.9.78.123
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.5
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
filelock==3.25.2
finnhub-python==2.4.27
firecrawl-py==4.22.0
flatbuffers==25.12.19
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-core==1.2.26
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.4
propcache==0.4.1
protobuf==6.33.6
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyrate-limiter==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-cache==1.3.1
requests-oauthlib==2.0.0
requests-ratelimiter==0.9.3
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
ta==0.11.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
url-normalize==2.2.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
zstandard==0.25.0
